feat: support listing mcrs
