+++ docs/megaport-cli_mcr_list_docs.md
@@ -1,0 +1,23 @@
+# docs
+
+Show documentation for this command
+
+## Description
+
+Display formatted documentation for this command from markdown files
+
+## Usage
+
+```sh
+megaport-cli mcr list docs [flags]
+```
+
+
+## Parent Command
+
+* [megaport-cli mcr list](megaport-cli_mcr_list.md)
+## Flags
+
+| Name | Shorthand | Default | Description | Required |
+|------|-----------|---------|-------------|----------|
+
